Replace old "use_directory_urls" config option with new "url_format".
Replace the old `use_directory_urls` boolean flag with a tri-state
option `url_format` which may be `directory`/`index` or `file`.  This
has the following effect:

    dir/file.md

Will be mapped to the following URL:
* [directory] -> `/dir/file/`
* [index] -> `/dir/file/index.html`
* [file] -> `/dir/file.html`

To retain backwards compatibility, the default is `directory` which is
the same behavior as the old default `use_directory_urls` = `True`.

This is respected under "serve" and "build" modes.  I have also updated
the unit tests and the Configuration section in the User Guide.