Fixed tests by moving mkdocs/test.py to test.py
If you run Python from inside the mkdocs package, it
doesn’t think mkdocs is a package and relative imports
are an error.

If you run Python from _outside_ the mkdocs package,
then it’s a package and you need relative imports.

test.py isn’t used by anything else in mkdocs, so the
simplest answer seems to be moving it outside the
package.