Windows: Use posixpath in nav.FC.make_absolute()
nav.FileContext.make_absolute() is only used to compare
links to entries in mkdocs.yml, so it makes sense for it
to only use POSIX-style paths.

Without this, the site fails to build because it checks for
a Windows-style path in a list of POSIX paths in
build.PathToURL.__call__().

This is the last change required to build the mkdocs site
on Windows.