Fix #183: Don't add classes to code blocks
Previously, MkDocs will add classes to the <pre><code> blocks generated
from the md file to allow for the prettify library to work its magic.
This commit moves the translation into base.js under mkdocs, so now the
theme handles the prettify instead of assuming in the build process.
This allows users to use other libraries, such as highlight.js.