Implement #163: Reload webpage when editing docs
This is the first pass at this feature request -
when running mkdocs using `mkdocs serve`, changes to the documentation
will cause a reload of the webpage, essentially live feedback.
We make use of python-livereload, a convenient wrapper around
LiveReload.
For this to work you need to install the browser extensions for your
browser.

The current implementation relies on watchdog to watch for changes in
the files in the `docs_dir`, which will then build to `site_dir`.
This will trigger the webpage reload because we have python-reload
watching `site_dir`.
A downside to this is that the static server gets a bit noisy when
serving, because when mkdocs builds the documentation, triggering the
python-livereload's handlers too many times, causing a lot of 'Ignore'
messages to be generated.