Update base.html
d0ugal marked #324 closed because of something to do with with a previous discussion about local store. That said: this templates src URLs are still "broken" (a bug) for non-HTTP-served content. 

It is just a reference error that only shows up when "http:" is not the default protocol to be handled -- i.e. when not being served. An alternative to this 'fix': store jquery and modernizr within the template directory structure itself and update this path reference to use {{base_url}}.