fixes test_copying_media
The test test_copying_media only worked in the first place,
because we always deliberately inserted the TOC tag and TOC delimiter
into markdown_source in convert_markdown. That way, the input passed
to md.convert was never empty.
Having removed the clutter of pre- and postprocessing the markdown_source
to insert the TOC tag, the input passed to md.covert was empty for the
first time, leading to the error
"AttributeError: 'Markdown' object has no attribute 'Meta'"

The test now creates a non-empty index.md file.