add documentation to sdist
When creating packages or installing from source it's good to have the
option of using the documentation provided with that code base.

I'm not sure if I've included the correct scope of items (really wasn't
sure about docs/CNAME) so if I need to edit this commit; please, let me
know.