Configuration Refactor
This change breaks up the config parsing into a set of validators. This
makes it somewhat easier to follow and add extra checks. As a result we
now also catch a few config errors we wouldn't have before.

The code probably still needs some tidying up, but it works well.