MkDocs Theme: search.html => search-modal.html
The `search.html` template name is a special name (hardcoded) which gets
written to the output if/when the theme includes a seperate search
page. However, the MkDocs theme uses a search modal instead. That does
not belong in a separate file in the output (it is included in `base.html`
template). By renaming the file, we still get the current fucntionality
(easy for users to override/replace the modal) but avoid generating a
broken `search.html` page in the output. Users can still create their own
standalone `search.html` file in the `theme_dir` is they want one, but now
they don't also break the modal when doing so.