Respect order of config extensions
In some particular case we need to reorder of builtin extensions,
and all the markdown extensions are putted at the `_end`. And the
builtin extensior are execute before our extensions.

With this change if a builtin extension is defined in the config
then is removed from builtin and added in the configuration order

In particular we need the `toc` extension after our `i18n`
extension to get the menu translated.