Ensure nav.homepage.url is not blank on error pages.
Also maintain pattern of ending urls with a slash in nav (all non-error pages
do so the error page shoudl also). A couple extra checks to ensure base_url
doesn't end in a slash though as it is expected that the string appended
to base_url will always begin with the slash.

Fixes #1131