Cleanup and refactoring of new structure
Some notes regarding various decisions made here:

And paths which start with a slash are assumed to be relative to the
docs_dir root. This behavior fixes #192. However, the slash not being
present in the output may surprise some users who are trying to create a
link relative to the server root when the mkdocs root is not at the
server root.

Still need to address links to files from the theme. I note that in the
homepage of our docs we link to `/img/favicon.ico`, which is a file
provided by the theme. It appears that the previous behavior was to
treat links which started with a slash as not local links and therefore
a check was not necessary. This then provided a workaround to avoid the
warning when linking to theme provided files.

* Page.url is the url relative to the site_dir
* Page.canonical_url is the relative url joined with site_url or None if
  site_url is not defined (the default).
* Page.abs_url is the path component of the canonical url or None if
  canonical_url is None.

Note that new behavior is slightly different than before. Previously
abs_url ignored site_url and was always url with '' prepended. With the
new behavior, is site_url includes a subdir, that subdir will be
included in the abs_url.

When not on a server, there is no sensable "absolute_url" for a page.
Therefore, we shouldn't try to define one.

The thinking is that users generating docs to be browsed in the local
file system (`file://`) should leave the site_url setting unset, while
users who will be serving their docs from a server should be setting the
site_url. And if the site_url point sot a subdir of the server, the
abs_url will stil be absolute from the server root as it uses the "path"
of the canonical_url of the page.

Note that without the magical url context all URLs must be prepended by
`{{ base_url }}/` in the templates. While this requires a change in
third party themes, it is more consistent.

The rest of the tests should probably be updated to use this.

Also load nav from files in get_navigation() rather than in config
validation. That way we only walk docs_dir once.

Note that the two failing tests need to be moved to the nav_tests to
ensure get_navigation() returns the proper nav structure when no nav
config is provided and the file structure is relied upon.

And add `files` keyword to `nav` event. Previously pages and nav were
synonymous. As that is no longer the case, a plugin may need access to
the files collection to properly manipulate the nav.