Refactor .travis.yml to use matrix. (#1275)
Travis-CI no longer installs all supported Python versions by default. We
need to define the Python version(s) we want in the `python` config setting
to ensure that that version is installed before running the tests.
Therefore, we need a matrix of `python` and `env` settings to ensure the
proper Python versions are matched up with their corresponding tox
environments. We can also isolate dependencies to specific envs.