Remove base_url handling from worker.
This makes for more consistent handling. Sometimes worker.js is run in a
seperate thread. Somethimes in the main thread (as a fallback) and
sometimes the index is prebuilt. That's three different envionments
where the index can be build, each with a different possable base_url.
By only applying the base_url in the main thread when search results are
displayed, we ensure consistancy.