Ensure paths in config are all unicode all the time.
When a byte string and a unicode string are both passed to os.path (join
ect) then returned value is always a byte string. Therefore, we need
every path string to be unicode. This ensures valudation checks that and
if the byte string uses the file system encoding, decodes it. For any
other encoding, a validation error is raised.