Add a url filter to templates.
Currently all local URLs must be prepended with base_url while external 
links must not. However, we can have both types of links in the same 
collection. As prepreprocessing of extra_css and extra_javascript 
settings is deprecated, we need a way to indicate the URL should be 
prepended, but only if it needs to be. The same applies to external 
links included in the site navigation. 

Therefore, a new temaplte filter is provided which any URL can be passed 
to. If the URL is absolute, it is returned unaltered. If it is relative 
and the template context includes a `page` object, then a URL is 
returned reltive to the page object. Otherwise, the URL is returned with 
base_url prepended. This also results in cleaner output as paths are 
normailized (redundant . are removed). Of course, the base_url tempalte 
variable remains (it is not being deprecated -- in fact the new filter 
relies on its existance) and template authors may continue to use it if 
they desire.