Add include_homepage_in_sidebar setting.
In the upstream Sphinx theme, this is controled when defining the 
toctree. However, Sphinx has a seperate settign for identifying the 
"homepage". In MkDocs the homepage is identified at the first item in 
the `nav`. Therefore, without this setting you can't exclude the 
homepage from the sidebar like you can in Sphinx. With this setting we 
get feature parity.