Warn on relative paths in nav.
Note that prior to 1.0 this was handled as part of the config 
validation. Now that it is moved out of config validation, we no longer 
have the option to collect all wartnings and then raise an error on 
strict mode. We now exist on the first error and never get to any later 
errors.

Closes #1604.