Cleaned up supported Python versions.
I updated the bash example which shows Python versions. Turns out those showed an old, unsupported version.

I didn't notice a second line was added for v0.x.x in the first commit. We offer no support for any version <1.0. Including <1.0 in the docs implies that we do offer some kind of support. Therefore, that line has been removed. To clarify, when I say we offer no support, I mean that bug fixes will not be backported. And there are various bugs which have been fixed in the 1.0 series that have not been backported and will never be. Therefore, it would be unfair to users to indicate that we still support those versions.

Finally, using `3.4+` indicates any future version (3.8, 3.9...). While we don't anticipate any problems, those are unknowns at this point in time. We shouldn't make promises we haven't (or can't) verify are true. Therefore, the hard list of versions as been restored.