Remove basepython setting from tox config
Travis has changed their default Python version from 2.7 to 3.6 on April 16, 2019 in anticipation of PY27 support being dropped in 2020. The `basepython` setting isn't nessecary as the tests will run by default in whichever version tox is called from. As we can be sure that version exists on the platform, it is best to use that version and avoids the need for future changes.