Update Travis config for default Python 2.7=>3.6 (#1782)
Travis has changed their default Python version from 2.7 to 3.6 on April
16, 2019 in anticipation of PY27 support being dropped in 2020. 

We are now running flake8 on Python 3.

Unfortunately, LinkChecker does not yet support Python 3. We need to
specify Python 2.7 in both the travis and tox configs to get the test to 
pass. We'll remove the 2.7 requirement as soon as LinkChecker is updated.