Initial implementation of local search fix
This is a first shot at fixing #871. It works by adding a config option to the search plugin
which generates an additional search_index.js file. This is simply search_index.json encapsulated in a
global variable. This gets embedded into the site, along with a small JavaScript shim,
by using the extra_javascript directive.