Fix url-filter with file urls on an empty site_url, closes #2189
Previously when "use_directory_urls" was false and "site_url" empty (e.g.
documentation for fs), the link to the home-page / main index was broken
missing /index.html at the end of the url.

Fix is to add the missing index.html in such a case by the url-template-
filter ( {{ ... | url }} ).

NOTE: Some values to the filter are undefined behavior as the filter
      will return wrong urls. These cases are considered theoretical
      (a file-name must end with a dot and be at the end of value) and not
      likely to happen.
      Declaring them undefined now, allows to change the implementation
      later in a compatible way.