Don't check the filesystem for edits to template files during a build
MkDocs templates are quite heavily split into files and can even be nested [1] (particularly for navs, which also repeat on every page).

Any time a template is referenced, even though it's cached in memory, Jinja ends up doing a filesystem access, to check whether the template file has been modified and so would need to be recompiled.
But it's not even useful to be able to modify a template file on disk in the middle of a mkdocs build.

So, by disabling autoreload, we eliminate all these filesystem accesses (on the order of thousands), other than the initial one per template.

[1]: https://github.com/mkdocs/mkdocs/blob/master/mkdocs/themes/mkdocs/nav-sub.html