Add warning about CNAME mismatch in gh-deploy
Although documentation was added in #1496 around needing to place the CNAME file for custom domains in your docs_dir, GitHub's default behavior of adding the CNAME directly to the gh-pages branch means that lots of folks will get bitten. This has the unfortunate and severe side effect of taking their site down and it's not immediately obvious what went wrong. This PR adds a new check: if there's a CNAME file in the gh-pages branch and it doesn't match the CNAME file in docs_dir (or the CNAME file in docs_dir doesn't exist) it will terminate the deployment. It adds a flag called "--ignore-cname" to bypass this check.

This is my first time contributing to mkdocs, so I hope I did everything as expected. Please let me know if there's any extra steps or stylistic changes needed.