No exception if directory already removed, closes #2420
When `mkdocs serve` is shutting down, it cleans up after itself removing
the temporary directory created earlier.

Previously the condition whether the directory is to be removed was
unchecked and prone to a file system race condition (shared resource).

Given the directory is absent on the file-system while shutting down,
`mkdocs serve` threw an exception and exited in failure.

Change is to override the cause of the exception by preventing the
attempt to remove a directory that is already gone.