draft: restore re-building with symbolic-links, closes #2425
previously (1.1.2 + master at 23e2051) building was triggered by changes
of file-content that was symbolically linked within docs_dir while
`mkdocs serve` was running.

since migrating from livereload>=2.6.1 to watchdog>=2.0.0 to detect
file-system changes (triggering the re-build) it stopped working.

this is because watchdog does not support symbolic links out of the box,
e.g. see [1].

change is to provide additional observe instructions on the realpath [2]
for the following cases:

1. docs_dir & config_file_path path deviation:

   when the absolute path to either the `docs_dir` or the `config_file` is
   different from its realpath, the realpath is added for observing (as
   well).

2. symbolic links within docs_dir:

   if a file within docs_dir is a symbolic link, the files real path
   is added for observing. sub-directories (that are not symbolically
   linked) are traversed up to a depth of nine levels (only if the
   recursive flag is enabled, otherwise no traversal into sub-directories).

Ref: 23e205153f01d24d50fe9ba18e5186cdbc2c2dbe
[1]: https://github.com/gorakhargosh/watchdog/issues/365
[2]: <https://docs.python.org/3.8/library/os.path.html#os.path.realpath>