Rework ConfigOption schemas as class-based
This is NOT a breaking change, the old style keeps working.

Now developers can make a subclass of Config, declare the schema of the config as fields of the class, and instances of this class will hold the processed config.

This better represents the relationship between what a config definition and a config instance is, now you think of configs definitions as classes and parsed configs as instances.

We also can write these fields as descriptors and enable safe attribute-based access. Static analysis will be able to see when a missing fields is accessed. And in followup changes I plan to add type annotations which will make even type checking fully sound.