Revert to pre-1.5 title rendering behavior
MkDocs 1.5 changed so that title detection is based fully on Markdown parsing, which makes it more reliable. But that also brought in another change - that Markdown actually gets partly rendered.

This way we can keep *only* the H1 tag detection and yank the title as soon as possible - before processing any inline elements.