Fix issue with handling `</>`
Markdown recently made a fix for an issue handling `</>` which would get
stripped out of content. Reference:
https://github.com/Python-Markdown/markdown/commit/64a3c0fbc00327fbfee1fd6b44da0e5453287fe4

MkDocs uses Markdown's monkey-patched HTMLParser, but does not use the
HTML extractor that has the logic to handle the things that have been
monkey-patched in. Because of this, MkDocs now fails when parsing
`</>`. This fixes the issue by adding the same logic to handle the issue
that Markdown uses.